src/requirements_rpi4.txt
# Raspberry Pi 4 (64-bit Raspberry OS / aarch64) requirements
# for src/main.py
#
# 설치 방법:
#   pip install -r requirements_rpi4.txt
#
# LCD 디스플레이(ST7789) 사용 시 추가 설치:
#   pip install st7789

# ──────────────────────────────────────────
# TFLite 추론
# 전체 TensorFlow 불필요 — aarch64 경량 런타임 사용
#
# tflite-runtime   : Python 3.9~3.11 지원 (구버전)
# ai-edge-litert   : tflite-runtime 후속 패키지
#                    Python 3.10~3.13 + aarch64 공식 지원 (v2.1.2+)
#
# Python 3.13에서는 imp 모듈 제거로 tflite-runtime 및 tensorflow가
# 동작하지 않으므로 ai-edge-litert 를 사용해야 합니다.
# ──────────────────────────────────────────
ai-edge-litert>=2.1.2

# ──────────────────────────────────────────
# 수치 연산
# ──────────────────────────────────────────
numpy>=1.24.0

# ──────────────────────────────────────────
# 오디오 feature 추출 (log-mel spectrogram)
# ──────────────────────────────────────────
librosa>=0.10.0

# ──────────────────────────────────────────
# 실시간 마이크 입력 / 스피커 출력
# ──────────────────────────────────────────
sounddevice>=0.4.6

# ──────────────────────────────────────────
# 오디오 파일 I/O (librosa 의존성)
# ──────────────────────────────────────────
soundfile>=0.12.1

# ──────────────────────────────────────────
# 신호 처리 (librosa 의존성)
# ──────────────────────────────────────────
scipy>=1.10.0

# ──────────────────────────────────────────
# LCD 이미지 처리 (ST7789 디스플레이)
# ──────────────────────────────────────────
Pillow>=9.0.0

# ──────────────────────────────────────────
# LLM 멘트 생성 (Gemini API)
# ──────────────────────────────────────────
google-genai>=1.0.0

# ──────────────────────────────────────────
# TTS 재생 (ElevenLabs API)
# MP3 재생을 위해 시스템 패키지도 필요:
#   sudo apt install mpg123
# ──────────────────────────────────────────
elevenlabs>=1.0.0

# ──────────────────────────────────────────
# HTTP 요청 (GPS / 날씨 / 대기질 API 조회)
# ip-api.com 및 Open-Meteo 엔드포인트 호출
# ──────────────────────────────────────────
requests>=2.28.0
